+++ back/main.py
@@ -1,12 +1,6 @@
 from pathlib import Path
 import re
-
-
-class PathNotFoundError(Exception):
-    def __init__(self):
-        msg = "Path dosn't exist"
-    def __str__(self):
-        return self.msg
+from exception import PathNotFoundError
 
 
 def searchSorting (addrs): 
@@ -24,32 +18,31 @@ def witre_on_txt(filetext):
             file.write(str(i)+"\n")
     return "done"
 
-# pathInput = input("Please Enter your Path: ")
-
-icon_adr = getaddress('/Users/SMD/Desktop/GW2/MaktabsharifMiniProject3/front/images/')
-html_adr = getaddress('/Users/SMD/Desktop/GW2/MaktabsharifMiniProject3/front')
-sorted_list = searchSorting(icon_adr)
-icon = icon_adr.glob("*icon*")
-html_files = html_adr.glob(".html")
-for i in html_files:
-    print(i)
-print(html_files)
-print(witre_on_txt(icon))
-
 def replace_icon_dir(html):
-    file = '/Users/SMD/Desktop/GW2/MaktabsharifMiniProject3/front/index.html'
-    with open(file,"r") as f:
+    with open(html,"r") as f:
         data = f.readlines()
-
-    with open(file, "w") as file_data:
+    with open(html, "w") as file_data:
         pattern = re.compile('(images.*-icon.*png)')
         for i in data:
-
             if re.search(pattern,i): 
                 file_data.writelines(i.replace("images","icon"))
             else:
                 file_data.writelines(i)
             
+
+
+icon_adr = getaddress('/Users/SMD/Desktop/GW2/MaktabsharifMiniProject3/front/images/')
+html_adr = getaddress('/Users/SMD/Desktop/GW2/MaktabsharifMiniProject3/front')
+sorted_list = searchSorting(icon_adr)
+icon = icon_adr.glob("*icon*")
+html_files = html_adr.glob("*.html")
+
+for i in html_files:
+    replace_icon_dir(i)
+    print(f'{i} has been repalced')
+
+print(witre_on_txt(icon))
+
