+++ front/contact.html
@@ -60,7 +60,7 @@
                         <a class="nav-link" href="#">Login</a>
                      </li>
                      <li class="nav-item">
-                        <a class="nav-link" href="#"><img src="images/serach-icon.png"></a>
+                        <a class="nav-link" href="#"><img src="icon/serach-icon.png"></a>
                      </li>
                   </ul>
                </div>
@@ -124,9 +124,9 @@
             </div>
             <div class="contact_menu">
                <ul>
-                  <li><a href="#"><img src="images/call-icon.png"></a></li>
+                  <li><a href="#"><img src="icon/call-icon.png"></a></li>
                   <li><a href="#">Call : +01 1234567890</a></li>
-                  <li><a href="blog.html"><img src="images/mail-icon.png"></a></li>
+                  <li><a href="blog.html"><img src="icon/mail-icon.png"></a></li>
                   <li><a href="features.html">demo@gmail.com</a></li>
                </ul>
             </div>
