+++ front/features.html
@@ -60,7 +60,7 @@
                         <a class="nav-link" href="#">Login</a>
                      </li>
                      <li class="nav-item">
-                        <a class="nav-link" href="#"><img src="images/serach-icon.png"></a>
+                        <a class="nav-link" href="#"><img src="icon/serach-icon.png"></a>
                      </li>
                   </ul>
                </div>
@@ -73,16 +73,16 @@
             <div class="row">
                <div class="col-lg-8 col-sm-12">
                   <div class="about_img"><img src="images/img-10.png"></div>
-                  <div class="like_icon"><img src="images/like-icon.png"></div>
+                  <div class="like_icon"><img src="icon/like-icon.png"></div>
                   <p class="post_text">Post By : 09/06/2019</p>
                   <h2 class="most_text">Most Awesome Blue Lake With Snow <br>foreste</h2>
                   <p class="lorem_text">Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis</p>
                   <div class="social_icon_main">
                      <div class="social_icon">
                         <ul>
-                           <li><a href="#"><img src="images/fb-icon.png"></a></li>
-                           <li><a href="#"><img src="images/twitter-icon.png"></a></li>
-                           <li><a href="#"><img src="images/instagram-icon.png"></a></li>
+                           <li><a href="#"><img src="icon/fb-icon.png"></a></li>
+                           <li><a href="#"><img src="icon/twitter-icon.png"></a></li>
+                           <li><a href="#"><img src="icon/instagram-icon.png"></a></li>
                         </ul>
                      </div>
                      <div class="read_bt"><a href="#">Read More</a></div>
@@ -130,9 +130,9 @@
             </div>
             <div class="contact_menu">
                <ul>
-                  <li><a href="#"><img src="images/call-icon.png"></a></li>
+                  <li><a href="#"><img src="icon/call-icon.png"></a></li>
                   <li><a href="#">Call : +01 1234567890</a></li>
-                  <li><a href="blog.html"><img src="images/mail-icon.png"></a></li>
+                  <li><a href="blog.html"><img src="icon/mail-icon.png"></a></li>
                   <li><a href="features.html">demo@gmail.com</a></li>
                </ul>
             </div>
